requirements.txt
feedparser
Misskey.py
python-dotenv
schedule
pip-review